FS-238 Remove F from Module
We refactor the `@module` macro so as to avoid the use of the `F[_]`
parameter, which is now added by the macro to the trait and vals.

As a smaller pass-by optimisations, we change the inheritance of the
market trait `ModuleLike` out of the object companion and add it to
the trait, so as to make the recursive test simpler.